+++ jsonargparse_tests/test_typehints.py
@@ -677,30 +677,34 @@ class TypeHintsTests(unittest.TestCase):
               class_path: {module}.Class
               init_args:
                   p1: 5
-                  __unresolved__:
-                    p2: '6'
-                    p3: 7.0
+              unresolved_init_args:
+                  p2: '6'
+                  p3: 7.0
             """
-            expected = Namespace(p1=5, p2='6', __unresolved__={'p3': 7.0})
+            expected = Namespace(
+                class_path=f'{module}.Class',
+                init_args=Namespace(p1=5, p2='6'),
+                unresolved_init_args={'p3': 7.0},
+            )
 
             parser = ArgumentParser(error_handler=None)
             parser.add_argument('--config', action=ActionConfigFile)
             parser.add_argument('--cls', type=Class)
 
             cfg = parser.parse_args([f'--config={config}'])
-            self.assertEqual(cfg.cls.init_args, expected)
+            self.assertEqual(cfg.cls, expected)
             cfg_init = parser.instantiate_classes(cfg)
             self.assertIsInstance(cfg_init.cls, Class)
-            self.assertEqual(cfg_init.cls.kwargs, expected.__unresolved__)
+            self.assertEqual(cfg_init.cls.kwargs, expected.unresolved_init_args)
 
-            cfg = parser.parse_args([f'--cls=Class', f'--cls.__unresolved__={expected.__unresolved__}'])
-            self.assertEqual(cfg.cls.init_args.__unresolved__, expected.__unresolved__)
+            cfg = parser.parse_args([f'--cls=Class', f'--cls.unresolved_init_args.p3=7.0'])
+            self.assertEqual(cfg.cls.unresolved_init_args, expected.unresolved_init_args)
 
             out = StringIO()
             with redirect_stdout(out), self.assertRaises(SystemExit):
                 parser.parse_args([f'--config={config}', '--print_config'])
-            init_args = yaml.safe_load(out.getvalue())['cls']['init_args']
-            self.assertEqual(init_args, expected.as_dict())
+            data = yaml.safe_load(out.getvalue())['cls']
+            self.assertEqual(data, expected.as_dict())
 
 
     def test_invalid_init_args_in_yaml(self):